requirements.txt
ppf.datamatrix
wand